requirement.txt
selenium==3.141.0
schedule==1.1.0
ddddocr==1.1.0
pyyaml==6.0
